ffado-mixer: Fix Python 3.10 type error in the crossbar router.
Thanks to Nils Philippsen for supplying this patch.  It fixes an int/float
type-related crash in the crossbar router associated with widget geometry
calculations.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2835 2be59082-3212-0410-8809-b0798e1608f0

+++ support/mixer-qt4/ffado/widgets/crossbarrouter.py
@@ -53,7 +53,7 @@ class VuMeter(QFrame):
         p = QPainter(self)
         value = self.level/4096
         r = self.rect()
-        r.setHeight(r.height() * math.sqrt(value))
+        r.setHeight(int(r.height() * math.sqrt(value)))
         r.moveBottom(self.rect().height())
         p.fillRect(r, self.palette().highlight())
